Improved check if file is locked

+++ source/private/Backup-EasitFolder.ps1
@@ -42,30 +42,56 @@ function Backup-EasitFolder {
             }
             Add-Type -Assembly System.IO.Compression.FileSystem
             $compressionLevel = [IO.Compression.CompressionLevel]::Optimal
+            $trySlowWay = $false
             if (Test-Path -Path $FolderToBackup -PathType Leaf) {
-                try {
-                    $zip = [System.IO.Compression.ZipFile]::Open($zipFile, 'create')
-                    $zip.Dispose()
-                    $zip = [System.IO.Compression.ZipFile]::Open($zipFile, 'update')
-                    [System.IO.Compression.ZipFileExtensions]::CreateEntryFromFile($zip, $FolderToBackup, $leaf, $compressionLevel) | Out-Null
-                    $zip.Dispose()
-                } catch {
-                    $zip.Dispose()
-                    Remove-Item $zipFile -Recurse -Force -Confirm:$false -InformationAction SilentlyContinue -ErrorAction SilentlyContinue
-                    Remove-Item $zip -Recurse -Force -Confirm:$false -InformationAction SilentlyContinue -ErrorAction SilentlyContinue
+                Write-Verbose "PathType is Leaf"
+                if (Test-ForFileLock -FilePath $FolderToBackup) {
                     $trySlowWay = $true
+                }
+                if (!($trySlowWay)) {
+                    try {
+                        $zip = [System.IO.Compression.ZipFile]::Open($zipFile, 'create')
+                        $zip.Dispose()
+                        $zip = [System.IO.Compression.ZipFile]::Open($zipFile, 'update')
+                        [System.IO.Compression.ZipFileExtensions]::CreateEntryFromFile($zip, $FolderToBackup, $leaf, $compressionLevel) | Out-Null
+                        $zip.Dispose()
+                    } catch {
+                        $zip.Dispose()
+                        Remove-Item $zipFile -Recurse -Force -Confirm:$false -InformationAction SilentlyContinue -ErrorAction SilentlyContinue
+                        Remove-Item $zip -Recurse -Force -Confirm:$false -InformationAction SilentlyContinue -ErrorAction SilentlyContinue
+                        $trySlowWay = $true
+                    }
                 }
             }
             if (Test-Path -Path $FolderToBackup -PathType Container) {
-                try {
-                    [IO.Compression.ZipFile]::CreateFromDirectory($FolderToBackup, $zipFile, $compressionLevel, $false)
-                } catch {
-                    Remove-Item $zipFile -Recurse -Force -Confirm:$false -InformationAction SilentlyContinue -ErrorAction SilentlyContinue
-                    $trySlowWay = $true
+                Write-Verbose "PathType is Container"
+                $filesToBackup = Get-ChildItem -Path $FolderToBackup -Recurse
+                if ($filesToBackup) {
+                    foreach ($fileToBackup in $filesToBackup) {
+                        if (Test-ForFileLock -FilePath $fileToBackup) {
+                            $trySlowWay = $true
+                        }
+                    }
                 }
+                if (!($trySlowWay)) {
+                    try {
+                        [IO.Compression.ZipFile]::CreateFromDirectory($FolderToBackup, $zipFile, $compressionLevel, $false)
+                    } catch {
+                        Remove-Item $zipFile -Recurse -Force -Confirm:$false -InformationAction SilentlyContinue -ErrorAction SilentlyContinue
+                        $trySlowWay = $true
+                    }
+                }
+                $filesToBackup = $null
             }
             if ($trySlowWay -or !(Test-Path -Path $zipFile)) {
-                Compress-Archive @compress -InformationAction SilentlyContinue
+                try {
+                    Write-Verbose "Trying to zip ($FolderToBackup) the old way"
+                    Compress-Archive @compress -ErrorAction Stop
+                } catch {
+                    Write-Warning "Unable to perform backup of $FolderToBackup"
+                    throw "$($_.Exception.Message)"
+                }
+                
             }
             Write-Verbose "Created zip archive for $FolderToBackup"
         } catch {
